Move env vars functions to config package
