+++ pkg/config/env_vars.go
@@ -1,4 +1,4 @@
-package utils
+package config
 
 import (
 	"log"
